requirements.txt
Django>=5.1.7
djangorestframework>=3.14.0
